+++ tensorflow_toolkit/image_retrieval/.gitignore
