+++ tensorflow_toolkit/image_retrieval/.pylintrc
