+++ tensorflow_toolkit/image_retrieval/README.md
