+++ tensorflow_toolkit/image_retrieval/configs/augmentation_config.json
