+++ tensorflow_toolkit/image_retrieval/image_retrieval/common.py
