+++ tensorflow_toolkit/image_retrieval/image_retrieval/frames_provider.py
@@ -14,7 +14,6 @@
 
 import os
 import cv2
-from textile.textile_cv_detector.detect_by_simple_dense_optical_flow import TextileDetector, get_rect_tl, get_rect_br
 
 
 class FramesProvider:
@@ -38,7 +37,7 @@ class FramesProvider:
 
 class CvatFramesProvider:
     def __init__(self, cvat_xmls_folder, videos_root_folder):
-        from textile.cvat_annotation import CvatAnnotation
+        from image_retrieval.cvat_annotation import CvatAnnotation
 
         self.should_go_to_next_video = False
         self.annotations = []
@@ -79,63 +78,5 @@ class CvatFramesProvider:
                     cv2.rectangle(view_frame, rect[0], rect[1], (255, 0, 255), 20)
 
                 frame_idx += 1
-
-                yield frame, probe_class, view_frame
-
-
-class VideoFramesProvider:
-    def __init__(self, filelist_path, videos_root_folder):
-        self.videos_list = []
-        self.videos_root_folder = videos_root_folder
-        self.frame_step = 5
-        self.should_go_to_next_video = False
-
-        data_dir = os.path.dirname(filelist_path)
-        with open(filelist_path) as f_list:
-            for line in f_list:
-                line = line.strip()
-                if line.startswith("#"):
-                    continue
-                chunks = line.split()
-                assert len(chunks) == 2
-                video_path, probe_class = chunks
-                self.videos_list.append({"rel_path": os.path.join(data_dir, video_path), "probe_class": probe_class})
-
-    def go_to_next_video(self):
-        self.should_go_to_next_video = True
-
-    def frames_gen(self):
-        for video_descr in self.videos_list:
-            rel_path = video_descr["rel_path"]
-            probe_class = video_descr["probe_class"]
-            video_path = rel_path # os.path.join(self.videos_root_folder, rel_path)
-
-            if not os.path.exists(video_path):
-                raise Exception('File not found: {}'.format(video_path))
-
-            textile_detector = TextileDetector(self.frame_step)
-
-            cap = cv2.VideoCapture(video_path)
-
-            while True:
-                if self.should_go_to_next_video:
-                    self.should_go_to_next_video = False
-                    break
-                _, frame = cap.read()
-
-                if frame is None:
-                    break
-
-                view_frame = frame.copy()
-
-                bbox = textile_detector.handle_frame(frame)
-                if bbox is not None:
-                    tl_x, tl_y = get_rect_tl(bbox)
-                    br_x, br_y = get_rect_br(bbox)
-
-                    frame = frame[tl_y:br_y, tl_x:br_x]
-                    cv2.rectangle(view_frame, (tl_x, tl_y), (br_x, br_y), (255, 0, 255), 20)
-                else:
-                    frame = None
 
                 yield frame, probe_class, view_frame
