+++ tensorflow_toolkit/image_retrieval/image_retrieval/losses.py
