+++ tensorflow_toolkit/image_retrieval/image_retrieval/model.py
