+++ tensorflow_toolkit/image_retrieval/requirements.txt
